Added Other Users profile page with delete post functionality

+++ backend/routes/post.js
@@ -28,8 +28,7 @@ router.post("/createpost", requirelogin, async (req, res) => {
 router.get("/fetchposts", requirelogin, async (req, res) => {
   try {
     const posts = await POST.find()
-      .populate("postedBy", "_id name userName")
-      .populate("comments.postedBy", "_id userName");
+      .populate("postedBy", "_id name userName").populate("comments.postedBy","_id userName")
     res.status(200).json(posts);
   } catch (error) {
     res.status(500).json({ error: "Internal Server Error" });
@@ -38,7 +37,7 @@ router.get("/fetchposts", requirelogin, async (req, res) => {
 
 router.get("/myposts", requirelogin, async (req, res) => {
   try {
-    const posts = await POST.find({ postedBy: req.user });
+    const posts = await POST.find({ postedBy: req.user }).populate("comments.postedBy", "_id userName").populate("postedBy", "_id userName") ;;
     res.status(200).json(posts);
   } catch (error) {
     res.status(500).json({ error: "Internal Server Error" });
@@ -105,7 +104,7 @@ router.put("/comment", requirelogin, async (req, res) => {
   }
 });
 
-router.put("/deletecomment", requirelogin, async (req, res) => {
+router.delete("/deletecomment", requirelogin, async (req, res) => {
   try{
     const newpost =await POST.findByIdAndUpdate(req.body.postId,{
       $pull:{comments:req.body.comm}
@@ -121,9 +120,12 @@ router.put("/deletecomment", requirelogin, async (req, res) => {
   }
 });
 
-router.put("/deletepost/:postId",requirelogin,async (req,res)=>{
+router.delete("/deletepost/:postId",requirelogin,async (req,res)=>{
   try{
     const result = await POST.findById(req.params.postId)
+    if(!result){
+      return res.status(422).json({msg:"Post Does not exist"})
+    }
     if(result.postedBy.toString()==req.user){
       const post = await POST.findByIdAndDelete(req.params.postId)
       res.status(200).json({post,msg:"Post Deleted successfully"})
